refactor(tests): updated tests to add node
